Add CI build, fix lints, ensure snapshots run with minimal tolerance so they work across platforms (#1)
* Add GHA CI build, fix lints, ensure snapshots only compared on Linux/x64

* Fix clippy lints

* Run snapshot tests on all plaftorm, introduce small tolerance

* Upload snapshot diffs on failure

+++ src/widgets/list.rs
@@ -233,7 +233,7 @@ impl List {
         if self.items.is_empty() {
             return;
         }
-        let visible = self.visible_rows() as i32;
+        let visible = self.visible_rows();
         let step = (visible - 1).max(1);
         self.move_selection(delta_pages * step);
     }
